Analysis: Suppress recycler view for emoji picker fragment on stories
This error is about not treat position as fixed; Instead, to only use
immediately and call holder.getAdapterPosition() to look it up later.

RecyclerView will not call onBindViewHolder again when the position of
the item changes in the data set unless the item itself is invalidated
or the new position cannot be determined.

For this reason, you should only use the position parameter while
acquiring the related data item inside this method, and should not keep
a copy of it.

If you need the position of an item later on (e.g. in a click listener),
use getAdapterPosition() which will have the updated adapter position.

However, and in order not to cause any unintended issues on the current
implementation of this screen, it is better to suppress this at the
moment, come back to it with a fix and add accompanying adequate
testing on top of the new solution afterwards.