Analysis: Disable icon location errors on app
The res/drawable folder is intended for density-independent graphics
such as shapes defined in XML. For bitmaps, move it to drawable-mdpi and
consider providing higher and lower resolution versions in
drawable-ldpi, drawable-hdpi and drawable-xhdpi. If the icon really is
density independent (for example a solid color) you can place it in
drawable-nodpi.

However, at this moment it is better to disable this warning as it only
applies to the 5 x 'intro0X.png' images.