Analysis: Suppress all deprecated warnings for mp4compose
Warning Messages:
- "'COLOR_FormatYUV420SemiPlanar: Int' is deprecated.
Deprecated in Java"
- "'INFO_OUTPUT_BUFFERS_CHANGED: Int' is deprecated.
Deprecated in Java"

These warnings are suppressed, that is, instead of them being resolved,
since a resolution would require a proper investigation and testing. As
such, it might be best to ignore this as out of scope, for now, and so
as to not introduce any breaking changes to this MP4 compose related
functionality overall.