Analysis: Resolve byte array type mismatch type warnings for mp4compose
Warning Message: "Type mismatch: inferred type is ByteArray? but
ByteArray was expected"

These calls to 'ByteArray' are not safe anyway, so it is better to add
the extra '!!' as this point, instead of hiding this crash via some
smart '?.let {.. }' alternative and end up with an app misbehavior to
this MP4 compose related functionality overall.