Analysis: Suppress listener deprecated warnings for photoeditor
Warning Message: "'onRemoveViewListener(Int): Unit' is deprecated. Use
{@link OnPhotoEditorListener#onRemoveViewListener(ViewType, int)}
instead"

These warnings are suppressed, that is, instead of them being resolved,
since a resolution would require a proper investigation and testing. As
such, it might be best to ignore this as out of scope, for now, and so
as to not introduce any breaking changes to this photo editor related
functionality overall.