Analysis: Suppress all deprecated warnings for photoeditor
Warning Messages:
- "'constructor MediaRecorder()' is deprecated.
Deprecated in Java"
- "'getter for defaultDisplay: Display!' is deprecated.
Deprecated in Java"
- "'getRealMetrics(DisplayMetrics!): Unit' is deprecated.
Deprecated in Java"
- "'get(Int): CamcorderProfile!' is deprecated.
Deprecated in Java"
- "'createCaptureSession((Mutable)List<Surface!>, CameraCaptureSession
StateCallback, Handler?): Unit' is deprecated. Deprecated in Java"
- "'AsyncTask<Params : Any!, Progress : Any!, Result : Any!>' is
deprecated. Deprecated in Java"
- "'THREAD_POOL_EXECUTOR: Executor!' is deprecated.
Deprecated in Java"
- "'setAudioStreamType(Int): Unit' is deprecated.
Deprecated in Java"
- "'constructor GestureDetector(GestureDetector.OnGestureListener)' is
deprecated. Deprecated in Java"
- "'constructor Handler()' is deprecated.
Deprecated in Java"

These warnings are suppressed, that is, instead of them being resolved,
since a resolution would require a proper investigation and testing. As
such, it might be best to ignore this as out of scope, for now, and so
as to not introduce any breaking changes to this photo editor related
functionality overall.