Analysis: Suppress all deprecated warnings for stories
Warning Messages:
- "'constructor Handler()' is deprecated.
Deprecated in Java"
- "'getSerializable(String?): Serializable?' is deprecated.
Deprecated in Java"
- "'getParcelableExtra(String!): T?' is deprecated.
Deprecated in Java"
- "'onBackPressed(): Unit' is deprecated. Overrides deprecated member in
'androidx.core.app.ComponentActivity'. Deprecated in Java"
- "'VIBRATOR_SERVICE: String' is deprecated.
Deprecated in Java"
- "'SYSTEM_UI_FLAG_*: Int' is deprecated.
Deprecated in Java"
- "'setter for systemUiVisibility: Int' is deprecated.
Deprecated in Java"
- "'FLAG_TRANSLUCENT_NAVIGATION: Int' is deprecated.
Deprecated in Java"
- "'setBottomSheetCallback(BottomSheetBehavior.BottomSheetCallback!):
Unit' is deprecated. Deprecated in Java"
- "'getColor(Int): Int' is deprecated.
Deprecated in Java"
- "'stopForeground(Boolean): Unit' is deprecated.
Deprecated in Java"
- "'getter for defaultDisplay: Display!' is deprecated.
Deprecated in Java"
- "'getSize(Point!): Unit' is deprecated.
Deprecated in Java"

These warnings are suppressed, that is, instead of them being resolved,
since a resolution would require a proper investigation and testing. As
such, it might be best to ignore this as out of scope, for now, and so
as to not introduce any breaking changes to this stories related
functionality overall.