Analysis: Suppress all deprecated warnings for app
Warning Messages:
- "'getter for systemWindowInsetTop: Int' is deprecated.
Deprecated in Java"
- "'startActivityForResult(Intent!, Int, Bundle?): Unit' is deprecated.
Deprecated in Java"
- "'FragmentPagerAdapter' is deprecated.
Deprecated in Java"

These warnings are suppressed, that is, instead of them being resolved,
since a resolution would require a proper investigation and testing. As
such, it might be best to ignore this as out of scope, for now, and so
as to not introduce any breaking changes to this app related module.