Build: Enable non transitive resources for the project
This commit doesn't updates the encrypted 'gradle.properties.enc' file
to point to that pinned hash which has the
'android.nonTransitiveRClass=true' property added into the 'gradle
.properties' file for the Loop project.

FYI: It seems that this is not necessary for the CI to pick-up this
specific change and download the correct version of 'gradle.properties'
to then build the project with non transitive resources enabled.
Instead, the CI is just copy-pasting the existing version of the
'gradle.properties-example' file, into the 'gradle.properties' file, and
then works itself from there.

PS: Maybe the 'gradle.properties.enc' file should be deleted from this
project in order to avoid further confusion on that matter going
forward.