Build: Disable jetifier
This commit updates the encrypted 'gradle.properties.enc' file to point
to that pinned hash which has the 'android.enableJetifier=false'
property added into the 'gradle.properties' file for the Loop project.

FYI: This is done in order for the CI to pick-up this specific change
and download the correct version of 'gradle.properties' to then build
the project with jetifier disable.